Check usage before delete location or hw-types

+++ console-web/webapp/src/main/java/org/openthinclient/web/thinclient/presenter/DirectoryObjectPanelPresenter.java
@@ -1,31 +1,43 @@
 package org.openthinclient.web.thinclient.presenter;
 
+import ch.qos.cal10n.IMessageConveyor;
+import ch.qos.cal10n.MessageConveyor;
 import com.vaadin.ui.*;
-import org.openthinclient.common.model.DirectoryObject;
-import org.openthinclient.common.model.Profile;
-import org.openthinclient.common.model.Realm;
+import org.openthinclient.common.model.*;
 import org.openthinclient.ldap.DirectoryException;
 import org.openthinclient.web.thinclient.ProfilePanel;
 import org.openthinclient.web.thinclient.ThinclientView;
+import org.openthinclient.web.thinclient.model.DeleteMandate;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.vaadin.viritin.button.MButton;
 
 import java.util.List;
-import java.util.Set;
+import java.util.function.Function;
+
+import static org.openthinclient.web.i18n.ConsoleWebMessages.*;
 
 /**
  * Presenter for DirectoryObjectPanel
  */
 public class DirectoryObjectPanelPresenter {
 
+  private static final Logger LOGGER = LoggerFactory.getLogger(DirectoryObjectPanelPresenter.class);
+
+  private final IMessageConveyor mc;
+
   ThinclientView thinclientView;
   ProfilePanel view;
   DirectoryObject directoryObject;
+  Function<DirectoryObject, DeleteMandate> deleteMandatSupplier;
 
   public DirectoryObjectPanelPresenter(ThinclientView thinclientView, ProfilePanel view, DirectoryObject directoryObject) {
 
     this.thinclientView = thinclientView;
     this.view = view;
     this.directoryObject = directoryObject;
+
+    mc = new MessageConveyor(UI.getCurrent().getLocale());
 
     view.getEditAction().addClickListener(this::handleEditAction);
     view.getDeleteProfileAction().addClickListener(this::handleDeleteAction);
@@ -51,31 +63,44 @@ public class DirectoryObjectPanelPresenter {
   public void handleDeleteAction(Button.ClickEvent event) {
 
     VerticalLayout content = new VerticalLayout();
-    Window window = new Window("Löschen bestätigen", content);
+    Window window = new Window(null, content);
     window.setModal(true);
     window.setPositionX(200);
     window.setPositionY(50);
 
-    content.addComponent(new Label("Wollen Sie das Profile " + directoryObject.getName() + " löschen?"));
-    HorizontalLayout hl = new HorizontalLayout();
-    hl.addComponents(new MButton("Cancel", event1 -> window.close()),
-                     new MButton("Löschen", event1 -> {
+    boolean deletionAllowed = true;
+    if (deleteMandatSupplier != null) {
+      DeleteMandate mandate = deleteMandatSupplier.apply(directoryObject);
+      deletionAllowed = mandate.checkDelete();
+      if (!deletionAllowed) {
+        window.setCaption(mc.getMessage(UI_COMMON_DELETE_NOT_POSSIBLE_HEADER));
+        content.addComponent(new Label(mandate.getMessage()));
+      }
+    }
 
-                       Realm realm = directoryObject.getRealm();
-                       try {
-                         realm.getDirectory().delete(directoryObject);
-                       } catch (DirectoryException e) {
-                         // TODO: handle exception
-                         // delete failed
-                         e.printStackTrace();
-                       }
+    if (deletionAllowed) {
+      window.setCaption(mc.getMessage(UI_COMMON_CONFIRM_DELETE));
+      content.addComponent(new Label(mc.getMessage(UI_COMMON_CONFIRM_DELETE_OBJECT_TEXT, directoryObject.getName())));
+      HorizontalLayout hl = new HorizontalLayout();
+      hl.addComponents(new MButton(mc.getMessage(UI_BUTTON_CANCEL), event1 -> window.close()),
+          new MButton(mc.getMessage(UI_COMMON_DELETE), event1 -> {
 
-                       // update display
-                       thinclientView.setItems(thinclientView.getAllItems());
-                       window.close();
-                       UI.getCurrent().removeWindow(window);
-                     }));
-    content.addComponent(hl);
+            Realm realm = directoryObject.getRealm();
+            try {
+              realm.getDirectory().delete(directoryObject);
+            } catch (DirectoryException e) {
+              // TODO: handle exception
+              // delete failed
+              e.printStackTrace();
+            }
+
+            // update display
+            thinclientView.setItems(thinclientView.getAllItems());
+            window.close();
+            UI.getCurrent().removeWindow(window);
+          }));
+      content.addComponent(hl);
+    }
 
     UI.getCurrent().addWindow(window);
 
@@ -99,5 +124,9 @@ public class DirectoryObjectPanelPresenter {
 
   public void hideEditButton() {
     view.getEditAction().setVisible(false);
+  }
+
+  public void setDeleteMandate(Function<DirectoryObject, DeleteMandate> deleteMandatSupplier) {
+    this.deleteMandatSupplier = deleteMandatSupplier;
   }
 }
